✨ feat: 팩토리 패턴을 통한 트러블슈팅 문서 생성 및 업데이트 로직 개선 (#27)
- 포스트 생성 시 필수 관계 설정 메서드 통합
- 중복 코드 제거 및 가독성 향상
